+++ js/main.js
@@ -1,5 +1,4 @@
-import introScreen from './screens/intro-screen';
+import introScreen from './screens/intro';
 import {showScreen} from './utils';
 
 showScreen(introScreen);
-
