+++ js/screens/game-2.js
@@ -1,17 +1,19 @@
-import statsScreen from './stats-screen';
+import game3Screen from './game-3';
 import * as utils from '../utils';
 
 const html = `<div class="game">
-    <p class="game__task">Найдите рисунок среди изображений</p>
-    <form class="game__content  game__content--triple">
+    <p class="game__task">Угадай, фото или рисунок?</p>
+    <form class="game__content  game__content--wide">
       <div class="game__option">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-      </div>
-      <div class="game__option  game__option--selected">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
-      </div>
-      <div class="game__option">
-        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+        <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+        <label class="game__answer  game__answer--photo">
+          <input name="question1" type="radio" value="photo">
+          <span>Фото</span>
+        </label>
+        <label class="game__answer  game__answer--wide  game__answer--paint">
+          <input name="question1" type="radio" value="paint">
+          <span>Рисунок</span>
+        </label>
       </div>
     </form>
     <div class="stats">
@@ -31,11 +33,11 @@ const html = `<div class="game">
   </div>`;
 
 const element = utils.getElementFromTemplate(html);
-const optionListNode = element.querySelectorAll(`.game__option`);
+const radioListNode = element.querySelectorAll(`input[type='radio']`);
 
-Array.from(optionListNode).forEach((item) => {
+Array.from(radioListNode).forEach((item) => {
   item.addEventListener(`click`, () => {
-    utils.showScreen(statsScreen, true);
+    utils.showScreen(game3Screen, true);
   });
 });
